requirements.txt
Flask==3.0.3
yt-dlp==2025.2.19
pywebview==5.4
